added helper function for splitting the tokens
