+++ src/tokenizer/token_list.c
@@ -41,4 +41,3 @@ void	token_addback(t_token **tokens, t_token *new)
 		tmp->next = new;
 	}
 }
-
